Split components into seperate files Add onClick handler for HeaderCell via new hook useSingleAndDoubleClick
