+++ src/components/MythicComponents/MythicResizableGrid/index.js
@@ -1,0 +1,3 @@
+import MythicResizableGrid from './MythicResizableGrid';
+
+export default MythicResizableGrid;
